Fixed an issue with identically named feeds
When two feeds in different categories had the same name all articles would be displayed in only the first feed. This resolves the issue. Note that feeds with the same name in the same category will still run into issues, as I currently don't see a need for this functionality.